+++ app/fast_api_app.py
@@ -173,13 +173,11 @@ else:
         pass
 
 
-from fastapi import Depends, FastAPI, HTTPException, Request
+from fastapi import FastAPI, HTTPException, Request
 from fastapi.middleware.cors import CORSMiddleware
 
 from app.services.sse_connection_limits import (
     SSERejectReason,
-    get_sse_connection_limit,
-    get_total_active_sse_count,
     release_sse_connection,
     try_acquire_sse_connection,
 )
@@ -237,9 +235,6 @@ else:
     # Populated by _init_adk_runners() from the lifespan background task.
     adk_app = None
     adk_app_fallback = None
-
-from app.app_utils.auth import verify_token
-from app.app_utils.typing import Feedback
 
 # Import persistent session service (with fallback to InMemory for local dev)
 if not BYPASS_IMPORT:
@@ -364,8 +359,6 @@ async def _init_adk_runners() -> None:
         logger.info("ADK runners initialized (background task complete)")
     except Exception as e:
         logger.exception("ADK runner initialization failed: %s", e)
-
-from datetime import datetime, timezone
 
 from slowapi.errors import RateLimitExceeded
 from slowapi.middleware import SlowAPIMiddleware
@@ -1036,6 +1029,7 @@ from app.routers.app_builder import router as app_builder_router
 from app.routers.approvals import router as approvals_router
 from app.routers.briefing import router as briefing_router
 from app.routers.byok import router as byok_router
+from app.routers.cache_admin import router as cache_admin_router
 from app.routers.community import router as community_router
 from app.routers.compliance import router as compliance_router
 from app.routers.configuration import router as configuration_router
@@ -1044,9 +1038,11 @@ from app.routers.data_io import router as data_io_router
 from app.routers.departments import router as departments_router
 from app.routers.document_viewer import router as document_viewer_router
 from app.routers.email_sequences import router as email_sequences_router
+from app.routers.feedback import router as feedback_router
 from app.routers.files import router as files_router
 from app.routers.finance import router as finance_router
 from app.routers.governance import router as governance_router
+from app.routers.health import router as health_router
 from app.routers.initiatives import router as initiatives_router
 from app.routers.integrations import router as integrations_router
 from app.routers.kpis import router as kpis_router
@@ -1123,576 +1119,9 @@ app.include_router(outbound_webhooks_router, tags=["Outbound Webhooks"])
 app.include_router(suggestions_router, tags=["Suggestions"])
 app.include_router(action_history_router, tags=["Action History"])
 app.include_router(recruitment_router)
-
-
-def _log_feedback_payload(payload: dict) -> None:
-    """Log feedback payload with compatibility across logger backends."""
-    if hasattr(logger, "log_struct"):
-        logger.log_struct(payload, severity="INFO")
-        return
-
-    import json as _json
-
-    logger.info("feedback=%s", _json.dumps(payload, default=str))
-
-
-@app.post("/feedback")
-def collect_feedback(feedback: Feedback) -> dict[str, str]:
-    """Collect and log feedback.
-
-    Args:
-        feedback: The feedback data to log
-
-    Returns:
-        Success message
-    """
-    _log_feedback_payload(feedback.model_dump())
-    return {"status": "success"}
-
-
-@app.get("/health/startup", tags=["Health"])
-async def health_startup():
-    """Startup probe for Cloud Run.
-
-    Checks that critical dependencies are reachable:
-    - Supabase connection
-    - Redis connection (if configured)
-    - Gemini credentials or Cloud Run ADC available
-
-    Returns 200 only when the app is ready to serve traffic.
-    Returns 503 if any critical dependency is unavailable.
-    """
-    checks: dict[str, str] = {}
-    all_ok = True
-
-    # Check Supabase
-    try:
-        from app.services.supabase import get_service_client
-
-        client = get_service_client()
-        # Simple query to verify connection
-        client.table("sessions").select("session_id").limit(1).execute()
-        checks["supabase"] = "ok"
-    except Exception as e:
-        checks["supabase"] = f"error: {e}"
-        all_ok = False
-
-    # Accept explicit credentials locally and Cloud Run ADC in GCP.
-    has_gemini_credentials = bool(
-        os.getenv("GOOGLE_API_KEY")
-        or os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
-        or (
-            os.getenv("GOOGLE_CLOUD_PROJECT")
-            and (os.getenv("K_SERVICE") or os.getenv("GOOGLE_CLOUD_RUN"))
-        )
-    )
-    checks["gemini_credentials"] = "ok" if has_gemini_credentials else "missing"
-    if not has_gemini_credentials:
-        all_ok = False
-
-    # Check Redis (non-critical — app degrades gracefully via circuit breaker)
-    try:
-        from app.services.cache import get_cache_service
-
-        cache = get_cache_service()
-        if cache._connected:
-            checks["redis"] = "ok"
-        else:
-            checks["redis"] = "not connected (non-critical)"
-    except Exception:
-        checks["redis"] = "not available (non-critical)"
-
-    status_code = 200 if all_ok else 503
-    return JSONResponse(
-        content={"status": "ready" if all_ok else "not_ready", "checks": checks},
-        status_code=status_code,
-    )
-
-
-def _health_response(
-    *,
-    status: str,
-    service: str,
-    latency_ms: int | None = None,
-    details: dict | None = None,
-    integrations: dict | None = None,
-) -> dict:
-    """Build canonical health response envelope (version 1).
-
-    All /health/* endpoints return this shape so parsers and monitors can
-    handle every endpoint uniformly without per-endpoint branching.
-
-    Args:
-        status: One of ``"ok"``, ``"degraded"``, or ``"down"``.
-        service: Short service identifier (e.g. ``"supabase"``, ``"redis"``).
-        latency_ms: Round-trip latency in milliseconds, or None if unmeasured.
-        details: Service-specific metadata dict.
-        integrations: Optional dict of third-party integration health entries.
-
-    Returns:
-        Canonical dict with ``status``, ``version``, ``service``,
-        ``latency_ms``, ``details``, ``checked_at``, and optionally
-        ``integrations``.
-    """
-
-    resp: dict = {
-        "status": status,
-        "version": "1",
-        "service": service,
-        "latency_ms": latency_ms,
-        "details": details or {},
-        "checked_at": datetime.now(timezone.utc).isoformat(),
-    }
-    if integrations is not None:
-        resp["integrations"] = integrations
-    return resp
-
-
-@app.get("/health/live", tags=["Health"])
-async def get_liveness():
-    """Fast liveness probe for container healthchecks.
-
-    Keep this endpoint dependency-free so Docker can quickly mark the
-    container healthy after restart even when downstream services are warming up.
-    """
-    return _health_response(status="ok", service="live", latency_ms=0)
-
-
-@app.get("/health/connections", tags=["Health"])
-async def get_connection_pool_health():
-    """Monitor Supabase connection pool stats and cache health."""
-    import time
-
-    required_env = ["SUPABASE_URL", "SUPABASE_SERVICE_ROLE_KEY"]
-    optional_critical_env = [
-        "WORKFLOW_STRICT_TOOL_RESOLUTION",
-        "WORKFLOW_ENFORCE_READINESS_GATE",
-        "BACKEND_API_URL",
-        "WORKFLOW_ALLOW_FALLBACK_SIMULATION",
-        "WORKFLOW_SERVICE_SECRET",
-    ]
-    missing_required_env = [k for k in required_env if not os.getenv(k)]
-    missing_critical_env = [k for k in optional_critical_env if not os.getenv(k)]
-    conn_start = time.monotonic()
-    try:
-        from app.rag.knowledge_vault import get_rag_client_stats, get_supabase_client
-        from app.services.supabase import get_client_stats, get_service_client
-        from app.services.supabase_async import execute_async
-
-        service_stats = get_client_stats()
-        rag_stats = get_rag_client_stats()
-
-        # Verify Service Client Connectivity
-        service_client: object
-        try:
-            service_client = get_service_client()
-            if not service_client:
-                raise ValueError("Service client failed to initialize")
-            # Lightweight connectivity check
-            await execute_async(
-                service_client.table("skills").select("count", count="exact").limit(0),
-                timeout=3.0,
-                op_name="health.connections.service_client",
-            )
-        except Exception as e:
-            raise ValueError(f"Service client connectivity check failed: {e}")
-
-        # Verify RAG Client Connectivity
-        try:
-            rag_client = await get_supabase_client()
-            if not rag_client:
-                raise ValueError("RAG client failed to initialize")
-            # Lightweight connectivity check — async client returns async builders
-            await (
-                rag_client.table("agent_knowledge")
-                .select("count", count="exact")
-                .limit(0)
-                .execute()
-            )
-        except Exception as e:
-            raise ValueError(f"RAG client connectivity check failed: {e}")
-
-        conn_latency_ms = int((time.monotonic() - conn_start) * 1000)
-
-        # Add Cache Health
-        from app.services.cache import get_cache_service
-
-        cache = get_cache_service()
-        cache_stats = await cache.get_stats()
-        cache_healthy = await cache.is_healthy()
-
-        cache_detail = {
-            "status": "healthy" if cache_healthy else "unhealthy",
-            "pool_max_connections": cache_stats.get("pool_max_connections"),
-            "latency_ms": cache_stats.get("latency_stats", {}),
-            "memory": cache_stats.get("memory_stats", {}),
-            "memory_alert": cache_stats.get("memory_stats", {}).get(
-                "memory_alert", False
-            ),
-            "circuit_breaker": cache_stats.get("circuit_breaker"),
-            "transport": "async_redis",
-        }
-
-        from app.services.supabase_resilience import supabase_circuit_breaker
-
-        supabase_cb = await supabase_circuit_breaker.get_status()
-
-        canary_raw = os.getenv("WORKFLOW_CANARY_USER_IDS", "")
-        canary_users = [u.strip() for u in canary_raw.split(",") if u.strip()]
-        workflow_rollout = {
-            "kill_switch_enabled": os.getenv("WORKFLOW_KILL_SWITCH", "false")
-            .strip()
-            .lower()
-            in {"1", "true", "yes", "on"},
-            "canary_enabled": os.getenv("WORKFLOW_CANARY_ENABLED", "false")
-            .strip()
-            .lower()
-            in {"1", "true", "yes", "on"},
-            "canary_user_count": len(canary_users),
-        }
-        config_readiness = {
-            "status": "ready" if not missing_required_env else "not_ready",
-            "missing_required": missing_required_env,
-            "missing_recommended": missing_critical_env,
-        }
-        if missing_required_env or missing_critical_env:
-            logger.warning(
-                "Configuration readiness issues detected. Missing required=%s missing_recommended=%s",
-                missing_required_env,
-                missing_critical_env,
-            )
-
-        # SSE connection stats (SSES-04 — observable from health endpoint)
-        try:
-            total_sse = await get_total_active_sse_count()
-        except Exception:
-            total_sse = None
-        sse_connections = {
-            "total_active": total_sse,
-            "per_user_limit": get_sse_connection_limit(),
-            "max_total": int(os.getenv("SSE_MAX_TOTAL_CONNECTIONS", "500")),
-        }
-
-        # Integration health from integration_sync_state (zero extra API calls)
-        integrations: dict = {}
-        try:
-            sc = get_service_client()
-            integrations_result = await execute_async(
-                sc.table("integration_sync_state").select(
-                    "provider, last_sync_status, last_sync_at, last_error_message"
-                ),
-                op_name="health.connections.integrations",
-            )
-            for row in integrations_result.data or []:
-                raw_status: str | None = row.get("last_sync_status")
-                if raw_status is None:
-                    int_status = "unknown"
-                elif any(kw in str(raw_status).lower() for kw in ("error", "failed")):
-                    int_status = "degraded"
-                else:
-                    int_status = "ok"
-                integrations[row["provider"]] = {
-                    "status": int_status,
-                    "last_sync_at": row.get("last_sync_at"),
-                }
-        except Exception as int_exc:
-            logger.warning(
-                "health.connections: could not load integration_sync_state: %s",
-                int_exc,
-            )
-
-        details = {
-            "pools": {"service_client": service_stats, "rag_client": rag_stats},
-            "efficiency_note": (
-                "Creation counts should remain stable (1) after initialization."
-            ),
-            "cache": cache_detail,
-            "supabase_circuit_breaker": supabase_cb,
-            "config_readiness": config_readiness,
-            "workflow_rollout": workflow_rollout,
-            "sse_connections": sse_connections,
-        }
-
-        return _health_response(
-            status="ok",
-            service="supabase",
-            latency_ms=conn_latency_ms,
-            details=details,
-            integrations=integrations if integrations else None,
-        )
-    except Exception as e:
-        logger.error(f"Health check failed: {e}")
-        return _health_response(
-            status="down",
-            service="supabase",
-            latency_ms=int((time.monotonic() - conn_start) * 1000),
-            details={"error": str(e)},
-        )
-
-
-@app.get("/health/workflows/readiness", tags=["Health"])
-async def get_workflow_readiness_health():
-    """Workflow preflight report for tool/integration readiness."""
-
-    try:
-        from app.workflows.readiness import build_workflow_readiness_report
-
-        report = build_workflow_readiness_report()
-        report["timestamp"] = datetime.now(timezone.utc).isoformat()
-        return report
-    except Exception as e:
-        logger.error(f"Workflow readiness health check failed: {e}")
-        return {
-            "status": "unhealthy",
-            "error": str(e),
-            "timestamp": datetime.now(timezone.utc).isoformat(),
-        }
-
-
-@app.get("/health/cache", tags=["Health"])
-async def get_cache_health():
-    """Monitor Redis cache health and performance with detailed diagnostics.
-
-    Returns:
-        - Connection status
-        - Circuit breaker state
-        - Connection pool statistics
-        - Cache hit/miss rates
-    """
-    import time
-
-    cache_start = time.monotonic()
-    try:
-        from app.services.cache import get_cache_service
-
-        cache = get_cache_service()
-
-        # Get basic stats
-        stats = await cache.get_stats()
-        is_healthy = await cache.is_healthy()
-        circuit_breaker = cache.get_circuit_breaker_state()
-        cache_latency_ms = int((time.monotonic() - cache_start) * 1000)
-
-        # Determine canonical status
-        cb_open = (
-            isinstance(circuit_breaker, dict) and circuit_breaker.get("state") == "open"
-        )
-        if is_healthy and not cb_open:
-            cache_status = "ok"
-        elif not is_healthy and cb_open:
-            # Circuit open but app still running — degraded, not down
-            cache_status = "degraded"
-        else:
-            cache_status = "ok" if is_healthy else "degraded"
-
-        details: dict = {
-            "connected": is_healthy,
-            "circuit_breaker": circuit_breaker,
-            "cache_stats": {
-                "hits": stats.get("hits", 0),
-                "misses": stats.get("misses", 0),
-                "hit_rate": stats.get("hit_rate", 0),
-            },
-        }
-
-        # Add connection info if available
-        if "redis_version" in stats:
-            details["connection"] = {
-                "redis_version": stats.get("redis_version"),
-                "used_memory": stats.get("used_memory_human"),
-                "connected_clients": stats.get("connected_clients"),
-            }
-
-        # Add error info if any
-        if "error" in stats:
-            details["error"] = stats.get("error")
-        if "reason" in stats:
-            details["reason"] = stats.get("reason")
-
-        return _health_response(
-            status=cache_status,
-            service="redis",
-            latency_ms=cache_latency_ms,
-            details=details,
-        )
-    except Exception as exc:
-        return _health_response(
-            status="down",
-            service="redis",
-            latency_ms=int((time.monotonic() - cache_start) * 1000),
-            details={"error": str(exc)},
-        )
-
-
-@app.get("/health/embeddings", tags=["Health"])
-async def get_embedding_health():
-    """Check Gemini embedding availability and latency."""
-    import time
-
-    emb_start = time.monotonic()
-    try:
-        from app.rag.embedding_service import get_embedding_health
-
-        health = get_embedding_health()
-        emb_latency_ms = int((time.monotonic() - emb_start) * 1000)
-
-        # Map old status values to canonical statuses
-        raw_status = health.get("status", "unhealthy")
-        emb_status = "ok" if raw_status == "healthy" else "down"
-
-        # The service already measures its own call latency; prefer it when present
-        service_latency = health.get("latency_ms")
-        latency = (
-            int(service_latency) if service_latency is not None else emb_latency_ms
-        )
-
-        details = {k: v for k, v in health.items() if k not in ("status", "latency_ms")}
-        return _health_response(
-            status=emb_status,
-            service="gemini",
-            latency_ms=latency,
-            details=details,
-        )
-    except Exception as exc:
-        return _health_response(
-            status="down",
-            service="gemini",
-            latency_ms=int((time.monotonic() - emb_start) * 1000),
-            details={"error": str(exc)},
-        )
-
-
-@app.get("/health/summarizer", tags=["Health"])
-async def get_summarizer_health() -> dict:
-    """Report conversation-summarizer rollout state.
-
-    Allows monitoring and operators to verify production state of the
-    ``ENABLE_CONVERSATION_SUMMARIZER`` flag and the effective
-    ``SESSION_MAX_EVENTS`` value without shelling into a container.
-
-    Returns:
-        Canonical health envelope with ``details`` carrying ``enabled`` (bool)
-        and ``session_max_events`` (int). Top-level ``enabled`` and
-        ``session_max_events`` are also mirrored for simple monitoring probes.
-    """
-    from app.persistence.supabase_session_service import (
-        ENABLE_CONVERSATION_SUMMARIZER as _summarizer_enabled,
-        SESSION_MAX_EVENTS as _session_max_events,
-    )
-
-    enabled = bool(_summarizer_enabled)
-    max_events = int(_session_max_events)
-    payload = _health_response(
-        status="ok",
-        service="summarizer",
-        latency_ms=0,
-        details={
-            "enabled": enabled,
-            "session_max_events": max_events,
-        },
-    )
-    # Mirror the two key fields at top-level so monitoring probes can read
-    # them without descending into ``details``.
-    payload["enabled"] = enabled
-    payload["session_max_events"] = max_events
-    return payload
-
-
-@app.get("/health/video", tags=["Health"])
-async def get_video_readiness():
-    """Check video generation configuration (Veo + Remotion). Read-only; no API calls."""
-    from app.services.video_readiness import get_video_readiness as get_readiness
-
-    report = get_readiness()
-
-    # Determine canonical status from configuration flags
-    veo_ok = report.get("veo_configured", False)
-    remotion_ok = report.get("remotion_configured", False)
-
-    if veo_ok and remotion_ok:
-        vid_status = "ok"
-    elif veo_ok or remotion_ok:
-        vid_status = "degraded"
-    else:
-        vid_status = "degraded"  # partial config — degraded, not down (app still runs)
-
-    return _health_response(
-        status=vid_status,
-        service="video",
-        latency_ms=0,
-        details=report.get("details", {}),
-    )
-
-
-def _csv_env(name: str) -> set[str]:
-    return {v.strip().lower() for v in os.getenv(name, "").split(",") if v.strip()}
-
-
-def _is_cache_admin(user: dict) -> bool:
-    """Authorizes access to admin cache invalidation endpoint.
-
-    Priority:
-    1) Explicit allow-any flag for controlled environments.
-    2) Explicit user allowlists (IDs/emails).
-    3) Role-based checks (default: admin, service_role).
-    """
-    if os.getenv("ALLOW_ANY_AUTH_ADMIN_ENDPOINT") == "1":
-        return True
-
-    user_id = str(user.get("id", "")).lower()
-    email = str(user.get("email", "")).lower()
-    allow_ids = _csv_env("ADMIN_USER_IDS")
-    allow_emails = _csv_env("ADMIN_USER_EMAILS")
-    allowed_roles = _csv_env("ADMIN_ROLES") or {"admin", "service_role"}
-
-    if user_id and user_id in allow_ids:
-        return True
-    if email and email in allow_emails:
-        return True
-
-    role_candidates = {str(user.get("role", "")).lower()}
-    metadata = user.get("metadata") or {}
-    if isinstance(metadata, dict):
-        app_meta = metadata.get("app_metadata") or {}
-        if isinstance(app_meta, dict):
-            app_role = app_meta.get("role")
-            if isinstance(app_role, str):
-                role_candidates.add(app_role.lower())
-            app_roles = app_meta.get("roles")
-            if isinstance(app_roles, list):
-                role_candidates.update(
-                    str(role).lower() for role in app_roles if isinstance(role, str)
-                )
-
-    return any(role in allowed_roles for role in role_candidates if role)
-
-
-@app.post("/admin/cache/invalidate", tags=["Admin"])
-async def invalidate_cache(
-    user_id: str | None = None,
-    confirm_flush_all: bool = False,
-    current_user: dict = Depends(verify_token),
-):
-    """Invalidate cache for a specific user or all users."""
-    if not _is_cache_admin(current_user):
-        raise HTTPException(status_code=403, detail="Admin privileges required")
-
-    from app.services.cache import get_cache_service
-
-    cache = get_cache_service()
-
-    if user_id:
-        await cache.invalidate_user_all(user_id)
-        return {"status": "success", "message": f"Cache invalidated for user {user_id}"}
-    else:
-        if not confirm_flush_all:
-            raise HTTPException(
-                status_code=400,
-                detail="confirm_flush_all=true is required for global cache invalidation",
-            )
-        # Invalidate all caches (use with caution)
-        await cache.flush_all()
-        return {"status": "success", "message": "All caches invalidated"}
+app.include_router(feedback_router)
+app.include_router(health_router)
+app.include_router(cache_admin_router)
 
 
 import asyncio
